Twill-less tool tests - handle repeats and conditionals.
Expand 'inputs' tree when building tool inputs for API functional tests.

Code is in there for both flat inputs (the way the UI would submit them), and a nested tree structure. The nested tree structure will not work with the Galaxy API currently - but it is the way the API should work :).