Utilize tool test runtime tree analysis earlier to more correctly determine which inputs are data params.
Turns out previous changesets attempting to separately disambiguate and supply implicit defaults for conditionals do not really work with data inputs - because they are not recognized as dataset input params during parsing. Additionally, this change should eliminate any ordering requirements of tool test params - i.e. you shouldn't need to provide the test parameters in an order that makes it clear what branch you are on before suppling a dataset in order for the parser to realize it is a dataset.

Finally, this changeset allows deletion of a lot of code and creates some more robust abstractions.